Improvement opportunities on deployment script
Improvement opportunity #1: Line 162. AGENT_ROLE was granted to the token address in the identityRegistry.
Change #1: Grant TOKEN_ROLE to the token address instead of the AGENT ROLE in the identityRegistry.

Improvement opportunity #2: Line 238 grants AGENT_ROLE to Token Agent in token. Line was already executed in line 100
Change #2: Delete line because it's repeated

Improvement opportunity #3: Line 243 grants AGENT_ROLE to Token Agent in identityRegistry. Line was already executed in line 161
Change #3: Delete line because it's repeated